Update react-router and refactor and combine Add/Edit user forms
